Fix tooltip bug when focusing spark line
Tab-focusing an element triggers a mouse event.

Before this change, that didn't work because we'd try to find the
cursor position, but there isn't one. (Specifically, `d3.pointer()`
threw an error.)

After this change, we handle both "regular" mouse events and focus
events. In the latter case, we put the tooltip next to the data point.